refactor: local storage contacts loading and saving for guest user

+++ src/app/services/contacts.service.ts
@@ -2,8 +2,7 @@ import { Injectable } from '@angular/core';
 import { Contact } from "../../models/contact";
 import { AuthService } from './auth.service';
 import { HttpClient } from '@angular/common/http';
-import { Router } from '@angular/router';
-import { Observable, Subject, from, lastValueFrom } from "rxjs";
+import { Subject, lastValueFrom } from "rxjs";
 import { environment } from "../../environments/environment.development";
 
 @Injectable({
@@ -24,7 +23,6 @@ export class ContactsService {
         this.authService.currentUser.loadLocalGuestContacts();
       } else {
         await this.syncRegisteredUserContacts();
-        console.log('contacts synced', this.authService.currentUser.contacts);
       }
       this.contacts$.next();
     }
@@ -49,12 +47,14 @@ export class ContactsService {
   async addContact(contact: Contact): Promise<Object | undefined> {
     if (this.authService.currentUser) {
       this.addContactLocally(contact);
-      const url = environment.BASE_URL + 'contacts';
-      const body = contact.toJson();
-      return lastValueFrom(this.http.post(url, body, {
-        headers: this.authService.getAuthTokenHeaders(),
-      }));
-    } else return;
+      if (this.authService.currentUser.user.id != 'guest') {
+        const url = environment.BASE_URL + 'contacts';
+        const body = contact.toJson();
+        return lastValueFrom(this.http.post(url, body, {
+          headers: this.authService.getAuthTokenHeaders(),
+        }));
+      }
+    } return;
   }
 
 
@@ -65,6 +65,7 @@ export class ContactsService {
       }
       this.authService.currentUser.contacts.push(contact);
       this.contacts$.next();
+      this.saveContactsIfGuest();
     }
   }
 
@@ -72,17 +73,14 @@ export class ContactsService {
   async updateContact(contact: Contact): Promise<Object | undefined> {
     if (contact.id != -1 && this.authService.currentUser) {
       this.updateContactLocally(contact);
-      if (this.authService.currentUser.user.id == 'guest') {
-        console.log('update contacts for guest user inactive so far');
-        return;
-      } else {
+      if (this.authService.currentUser.user.id != 'guest') {
         const url = environment.BASE_URL + 'contacts/' + contact.id;
         const body = contact.toJson();
         return lastValueFrom(this.http.put(url, body, {
           headers: this.authService.getAuthTokenHeaders(),
         }));
       }
-    } else return;
+    } return;
   }
 
 
@@ -92,6 +90,7 @@ export class ContactsService {
       if (contactsArrayIndex >= 0) {
         this.authService.currentUser.contacts[contactsArrayIndex] = contact;
         this.contacts$.next();
+        this.saveContactsIfGuest();
       } else {
         console.error('Task with ID', contact.id, 'could not be updated (task not found).');
       }
@@ -102,11 +101,13 @@ export class ContactsService {
   async deleteContact(id: number): Promise<Object | undefined> {
     if (id != -1 && this.authService.currentUser) {
       this.deleteContactLocally(id);
-      const url = environment.BASE_URL + 'contacts/' + id;
-      return lastValueFrom(this.http.delete(url, {
-        headers: this.authService.getAuthTokenHeaders(),
-      }));
-    } else return;
+      if (this.authService.currentUser.user.id != 'guest') {
+        const url = environment.BASE_URL + 'contacts/' + id;
+        return lastValueFrom(this.http.delete(url, {
+          headers: this.authService.getAuthTokenHeaders(),
+        }));
+      }
+    } return;
   }
 
 
@@ -116,9 +117,17 @@ export class ContactsService {
       if (contactsArrayIndex >= 0) {
         this.authService.currentUser.contacts.splice(contactsArrayIndex, 1);
         this.contacts$.next();
+        this.saveContactsIfGuest();
       } else {
         console.error('Task with ID', id, 'could not be deleted (task not found).');
       }
+    }
+  }
+
+
+  saveContactsIfGuest(): void {
+    if (this.authService.currentUser && this.authService.currentUser.user.id == 'guest') {
+      this.authService.currentUser.saveLocalGuestContacts();
     }
   }
 }
